backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
apscheduler==3.10.4
feedparser==6.0.11
aiohttp==3.11.11
pycoingecko==3.1.0
yfinance==1.2.0
praw==7.8.1
python-dotenv==1.0.1
aiohttp-socks==0.9.0
